Add code generation for binary operators
Corresponds to Chapter 6 ("Extending the Language: User-defined
Operators") from Kaleidoscope tutorial...

however, I was unable to run the Mandelbrot set example, because I can
no longer call the predefined extern subroutines.  This problem appears
to have begun in this repo when I moved to CMake, however, the problem
also appears in the provided source code for the chapter compiled with
a Makefile.

All the more reason to have testing in the future.